--- src/types/Survey/Survey.ts
@@ -1,8 +1,0 @@
-export interface SurveyState {
-  nickname: string;
-  gender: string;
-  age: number;
-  height: number;
-  weight: number;
-  split: number;
-}
